Make Gubernator github hook processing more memory-efficient.
Bail out entirely if it's unlikely we'll succeed.
